reorg rpc mod
